requirements.txt
pyspark
prometheus-client
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
